+++ packages/components/src/Tile/subcomponents/GenericTile/GenericTile.spec.stories.tsx
@@ -1,7 +1,6 @@
 import React from "react"
 import type { Meta, StoryObj } from "@storybook/react"
-import { expect, within, userEvent } from "@storybook/test"
-
+import { expect, within, waitFor } from "@storybook/test"
 import { GenericTile } from "./GenericTile"
 
 const meta: Meta<typeof GenericTile> = {
@@ -19,11 +18,72 @@ export default meta
 type Story = StoryObj<typeof GenericTile>
 
 export const Flip: Story = {
-  play: async ({ canvasElement }) => {
+  play: async ({ canvasElement, step }) => {
     const canvas = within(canvasElement)
 
-    await userEvent.click(canvas.getByRole("button", { name: "Information" }))
+    await step("initial render complete", async () => {
+      await waitFor(() => {
+        canvas.getByRole("button", {
+          name: "View more information: Title",
+        })
+      })
+    })
 
-    await expect(canvas.getByText("Side B")).toBeInTheDocument()
+    await step("Can focus to tile", async () => {
+      await waitFor(() => {
+        expect(canvas.getByText("Side B")).toBeInTheDocument()
+      })
+    })
+  },
+}
+
+export const InfoButtonLabelDefault: Story = {
+  play: async ({ canvasElement, step }) => {
+    const canvas = within(canvasElement)
+
+    await step("initial render complete", async () => {
+      await waitFor(() => {
+        canvas.getByRole("button", {
+          name: "View more information: Title",
+        })
+      })
+    })
+
+    await step("Can focus to button", async () => {
+      await waitFor(() => {
+        const buttonWithInfoLabel = canvas.getByRole("button", {
+          name: "View more information: Title",
+        })
+        buttonWithInfoLabel.focus()
+        expect(buttonWithInfoLabel).toHaveFocus()
+      })
+    })
+  },
+}
+
+export const InfoButtonLabel: Story = {
+  args: {
+    infoButtonLabel: "Test Label",
+  },
+  play: async ({ canvasElement, step }) => {
+    const canvas = within(canvasElement)
+
+    await step("initial render complete", async () => {
+      await waitFor(() => {
+        canvas.getByRole("button", {
+          name: "Test Label",
+        })
+      })
+    })
+
+    await step("Can focus to button", async () => {
+      await waitFor(() => {
+        const buttonWithInfoLabel = canvas.getByRole("button", {
+          name: "Test Label",
+        })
+        buttonWithInfoLabel.focus()
+        expect(buttonWithInfoLabel).toHaveFocus()
+      })
+    })
   },
 }